% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BCI.R
\name{BCI}
\alias{BCI}
\title{BCI}
\usage{
BCI(object, points = NA, AOU = NA, years = NA, times = NA, band = 1,
  intervals = NA, visits = NA, reps = NA, output = "dataframe", ...)
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a \code{list} of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{times}{A numeric vector. Returns data only from points where the years the point was visited  matches one of the values in \code{years} The year a visit takes place is determined by the \code{Year} column in the \code{visits} slot which is derived from the information in the \code{EventDate} column.}

\item{visits,}{The visits that will be used for the matrix. Defautls to \code{c(1,2)}.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}

\item{...}{Additional arguments passed to \code{getBirds}}

\item{band.}{A numeric vector. Defaults to 1. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}
}
\description{
Calcualtes the Bird Conservation Index (BCI)
}
\details{
This produces a Count X Visit matrix for a \code{NPSbirds} object or a \code{list} of such objects. Each row of the matrix will correspond to a different pointt in a different year. The columns of the matrix will be the park code, the point name, the year visited, and a column of abundances of the indcated species at that visit. If multiple species are indicated in \code{AOU}, their abundances will be totaled.
}

