% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCIMap.R
\name{BCIMap}
\alias{BCIMap}
\title{BCIMap}
\usage{
BCIMap(object, Dir, years = NA, points = NA, color = NA,
  APIkey = NA, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{years}{A vector of numbers that will return only data from the indicated years.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{color}{A character string naming a valid color for figures. Default is grayscale.}

\item{APIkey}{A character string of a Google API key used for generating maps with Google basemaps. If no APIkey is provided, basemaps default to stamen map.}

\item{...}{Additonal arguments passed to \code{\link{BCIMap}}.}
}
\description{
Generate a map of bird community index (BCI) scores at network and park unit levels.
}
\details{
This function calculates point-level bird community index (BCI) scores and generates a map showing BCI scores plotted using a color ramp.
}
\references{
O'Connell, T., R. Brooks, M. Lanzone, and J. Bishop. 2003. A bird community index for the mid-Atlantic piedmont and coastal plain, Final Report to the USGS-Patuxent Wildlife Research Center. Penn State Cooperative Wetlands Center, University Park, p 44. Report.
O'Connell, T. J., L. E. Jackson, and R. P. Brooks. 1998. The bird community index: A tool for assessing biotic integrity in the mid-Atlantic highlands. Penn State Cooperative Wetlands Center; EPA.
O'Connell, T. J., L. E. Jackson, and R. P. Brooks. 2000. Bird guilds as indicators of ecological condition in the central Appalachians. Ecological Applications 10:1706-1721.
Ladin, H., Z. S., and W. G. Shriver. 2016. Using regional bird community dynamics to evaluate ecological integrity within national parks. Ecosphere 7.
}
