% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCIPlot.R
\name{BCIPlot}
\alias{BCIPlot}
\title{BCIPlot}
\usage{
BCIPlot(
  object,
  years = NA,
  points = NA,
  visits = NA,
  times = NA,
  plot_title = NA,
  point_num = NA,
  type = "Cent_Appal",
  caption = T,
  palette = "BuGn",
  output = "total",
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{birdRichness()}.}

\item{years}{A numeric vector. Indicates which years should be graphed.}

\item{points}{A character vector of point names. Only these points will be used.}

\item{visits}{A length 1 numeric vector, defaults to NA. Returns data only from the indicated visits.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been vistied is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{plot_title}{Optional,  A title for the plot.}

\item{point_num}{An optional numeric vector indicating the number of points sampled each year. If \code{object} is a \code{NCRNbirds} object
then this will be calculated automatically. If \code{object} is a \code{data.frame} or a \code{list} than this can be provided by the user.}

\item{type}{A mandatory length 1 character \code{vector} that indicates the type of BCI to calculate. Can be "Cent_Appal" (default) 
or "NETN_Forest_BCI"}

\item{caption}{Logcial. Indicates if the default caption "Values in parentheses indicate the number of points monitored each visit of each year."
should be included.}

\item{palette}{Color pallete for the background of the graph. Defaults to "BuGn" (blue green) but will accept any RColorBrewer palette}

\item{output}{Either "total" (the default) or "list". Only used when \code{object} is a \code{list}.}

\item{plot}{Logical. Return plot \code{TRUE} (default) or data.frame \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link{birdRichness}}}
}
\description{
Plots BCI score over time.
}
\details{
This function produces a graph the Bird community Index  over time. It does this by using the output of 
the \code{\link{BCI}} function, and averging the plot values for the park's yearly value. The error bars on the plot represnt 95% 
bootstrap confidence intervals. The data is then passed on 
to ggplot2 for graphing. Typically this is done automatically by providing an \code{NCRNbirds} object or a \code{list}
of such objects. If the user wishes to provide their own \code{data.frame} it should have 3 columns, \code{Year, BCI, BCI_Category}
and each row should be data from single year.
}
