% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdGuildSummary.R
\name{BirdGuildSummary}
\alias{BirdGuildSummary}
\title{BirdGuildSummary}
\usage{
BirdGuildSummary(
  object,
  years = NA,
  points = NA,
  visits = NA,
  times = NA,
  output = "dataframe",
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{birdRichness()}.}

\item{years}{A numeric vector. Indicates which years should be summarized.}

\item{points}{A character vector of point names. Only these points will be used.}

\item{visits}{A length 1 numeric vector, defaults to NA. Returns data only from the indicated visits.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been vistied is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}

\item{...}{Additional arguments passed to \code{\link{BCI}}}
}
\value{
Returns a \code{data.frame} of the mean and SD of percent guild membership.
}
\description{
Calculates a summary of guild membership across point counts.
}
\details{
This calculates a summary of the percent guild membership across point counts.
}
\references{
O'Connell, TJ, LE Jackson and RP Brooks. 1998. The Bird Community Index: A Tool for Assessing Biotic Integrity in the 
Mid-Atlantic Highlands. Final Report prepared for U. S. Environmental Protection Agency, Region III.

O'Connell TJ, LE Jackson and RP Brooks. 2000. Bird guilds as indicators of ecological condition in the central Appalachians. 
Ecological Applications 10:1706-1721.
}
