% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSpeciesInfo.R
\name{CheckSpeciesInfo}
\alias{CheckSpeciesInfo}
\title{CheckSpeciesInfo}
\usage{
CheckSpeciesInfo(object, Dir, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{Dir}{A filepath directory to where the data resides.}

\item{...}{Additonal arguments passed to \code{\link{CheckSpeciesInfo}}.}
}
\description{
Query for comparing species information within raw data (e.g., Common Name, Scientific Name, and 4-letter Alpha coes), with up-to-date taxonomic information to standardize data prior to analysis.
}
\details{
This function uses resulting data gathered after running the \code{\link{getUpdatedSpeciesInfo}} function to query existing taxonomic information within raw data, and identifies any species that have discrepancies .
}
\references{
Chesser, R. T., K. J. Burns, C. Cicero, J. L. Dunn, A. W. Kratter, I. J. Lovette, P. C. Rasmussen, J. V. Remsen, Jr., D. F. Stotz, B. M. Winger, and K. Winker. 2018. Check-list of North American Birds (online). American Ornithological Society. http://checklist.aou.org/taxa

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson. 2018. North American Breeding Bird Survey Dataset 1966 - 2017, version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research Center. https://doi.org/10.5066/F76972V8.

Panjabi, A. O., P. J. Blancher, W. E. Easton, J. C. Stanton, D. W. Demarest, R. Dettmers, K. V. Rosenberg, P. in Flight Science Committee, and others. 2017. The partners in flight handbook on species assessment version 2017. Partners in Flight.
}
