% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovsXVisit.R
\name{CovsXVisit}
\alias{CovsXVisit}
\title{CovsXVisit}
\usage{
CovsXVisit(object, points = NA, AOU = NA, band = 1, visits = NA,
  years = NA, output = "dataframe", ...)
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a \code{list} of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{band}{}

\item{visits}{The visits that will be used for the matrix. Defautls to \code{NA}. See Details below}

\item{years}{A vector of numbers. will return only data from the indicated years.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}

\item{...}{Additional arguments passed to \code{getBirds}}
}
\description{
Produces a compoasite of Covariates X Visits matrix for use in analyses.
}
\details{
This generates and combines matrices of the following covariates: temperature, humidity, sky condition, wind, time of day, ordinal day, and observer by Visits for use in building occupancy and abundance models with  \code{estimateOccupancy} and \code{estimateAbundance} functions, respectively. If \code{visits} is left as \code{NA} then the visits used will be 1 through the number of visits indicated in the \code{visits} slot. Otherwise a numeric vectore e.g. c(1,2) can be used to select which visits are used.
}
