% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectsPlot.R
\name{detectsPlot}
\alias{detectsPlot}
\title{detectsPlot}
\usage{
detectsPlot(object, years = NA, points = NA, visits = NA, times = NA,
  plot_title = NA, point_num = NA, se = FALSE, output = "total", ...)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{CountXVisit()}.}

\item{years}{A numeric vector. Indicates which years should be graphed.}

\item{points}{A character vector of point names. Only these points will be used.}

\item{visits}{A length 1 numeric vector, defaults to NA. Returns data only from the indicated visits.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been vistied is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{plot_title}{Optional,  A title for the plot.}

\item{point_num}{An optional numeric vector indicating the number of points sampled each year. If \code{object} is a \code{NCRNbirds} object
or a \code{list} of such objects, then this will be calculated automatically. If \code{object} is a \code{data.frame} than this can be
provided by the user.}

\item{se}{Add Standard error to the plot. \code{TRUE} or \code{FALSE}, defaults to  \code{FALSE}}

\item{output}{Either "total" (the default) or "list". Only used when \code{object} is a \code{list}.}

\item{...}{Additional arguments passed to \code{\link{CountXVisit}}}
}
\description{
Plots bird detections (relative abundance) over time.
}
\details{
This function produces a graph of species detections over time. It does this by using the output of the \code{\link{CountXVisit}}
function. The data is then passed on to ggplot2 for graphing.
}
