% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAbundance.R
\name{estimateAbundance}
\alias{estimateAbundance}
\title{estimateAbundance}
\usage{
estimateAbundance(object, points = NA, AOU = NA, years = NA,
  times = NA, band = 1, visits = NA, site.covs = NA,
  obs.covs = NA, mixture = "P", Table = TRUE, Figure = TRUE,
  color = NA, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{times}{A numeric vector of length 1 passed on to \code{\link{CountXVisit}} and from there to \code{\link{getVisits}}. Returns only data from points where the number of years that a point has been visited is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{band}{A numeirc vector. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}

\item{visits}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{site.covs}{A list of desired site-level covariates that can be included within the abundance estimation process of the N-mixture model.}

\item{obs.covs}{A list of desired observation-level covariates that can be included within the detection probability estimation process of the N-mixture model.}

\item{mixture}{Defaults to "P". One of either "P", "NB", or "ZIP". Indicates which latent abundace distibution to use, either Poisson, Negative Binomial or Zero Inflated Poinsson.  Passed on to the \code{mixture} argumnet of \code{pcount} in the \code{unmarked} package.}

\item{Table}{Logical (default = TRUE). Print table of results to the console.}

\item{Figure}{Logical (default = TRUE). Show figure of trend in abundance over time.}

\item{color}{A character input of a valid color name for determining color in figure (e.g., "red").}

\item{...}{Additonal arguments passed to \code{\link{CountXVisit}} and from there to \code{\link{getBirds}}.}
}
\description{
Peforms trend analysis on bird data using N-mixture models (Royal 2004) form the unmarked package, and exports tables and figures.
}
\details{
This function is a wrapper for the \code{\link[NCRNbirds]{nmixBirds_zl}} for fitting a N-mixture model (Royal 2004, Kery et al. 2005) to data in an NCRNbirds object or a list of such objects. The data is first extracted from the object(s) and then fed to the \code{\link[unmarked]{pcount}} function. Tables and figures of resutls of the analysis are returned, and may be saved.
}
\references{
Kery, M. Royle, J.A. and Schmid. 2005. Modeling avian abundance from replicated counts using binomail mixutre models. Ecological Applications. 15: 1450-1461.

Royle, J. A.  2004. N-Mixture models for estimating population size from spatially replicated counts. Biometrics 60: 108-115.
}
