% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateOccupancy.R
\name{estimateOccupancy}
\alias{estimateOccupancy}
\title{estimateOccupancy}
\usage{
estimateOccupancy(object, points = NA, AOU = NA, years = NA,
  times = NA, band = 1, visits = NA, site.covs = NA,
  obs.covs = NA, Table = TRUE, Figure = TRUE, color = NA, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been visited is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{band}{A numeirc vector. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}

\item{visits}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{site.covs}{A list of desired site-level covariates that can be included within the estimation of probability of site occupancy.}

\item{obs.covs}{A list of desired observation-level covariates that can be included within the detection probability estimation process of the occupancy model.}

\item{Table}{Logical (default = TRUE). Print table of results to the console.}

\item{Figure}{Logical (default = TRUE). Show figure of trend in occupancy over time.}

\item{color}{A character input of for determining color in figure (e.g., "red").}

\item{...}{Additonal arguments passed to \code{estimateOccupancy}.}
}
\description{
Fit hierarchical occupancy models to count data (MacKenzie et al. 2002) with the \code{occu()} function (Fiske et al. 2011).
}
\details{
This function allows users to fit occupancy models and select site-level and observation-level covariates (MacKenzie et al. 2002) to data in an NCRNbirds object or a list of such objects. The data is first extracted from the object(s) and is then fed to the \code{\link[unmarked]{occu}} function and the resutls of the analysis are returned.
}
\references{
MacKenzie, D. I., J. D. Nichols, G. B. Lachman, S. Droege, J. Andrew Royle, and C. A. Langtimm. 2002. Estimating site occupancy rates when detection probabilities are less than one. Ecology 83:2248-2255.
 @references Fiske, I., R. Chandler, and others. 2011. Unmarked: An R package for fitting hierarchical models of wildlife occurrence and abundance. Journal of Statistical Software 43:1-23.
}
