% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBirdNames.R
\name{getBirdNames}
\alias{getBirdNames}
\title{getBirdNames}
\usage{
getBirdNames(object, names = NA, in.style = "AOU", out.style = "common")
}
\arguments{
\item{object}{A \code{NCRNbirds} object or a \code{list} of such objects.}

\item{names}{A character vector containing the names to be translated.}

\item{in.style}{A length 1 character vector. Either "AOU", "Latin" or "common".  Indicates the type of name found in the "names" agrument.}

\item{out.style}{A length 1 character vector. Either "AOU", "Latin", "common". Indicates the type of names to be returned. Defaults to "common"
# @param output Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}
}
\description{
Translates between Latin name, common name and AOU code based on data from the \code{Species} slot.
}
\details{
This function converts a vector of names between the American Ornitholgoical Union Code (AOU Code), Latin name and scientific name. This is used by various other functions to provide the correct type of name in the output. The data to make these conversion is stored in the \code{Species} solot of a \code{NCRNbirds} object. 
 
 Currently the \code{list} method simply uses the data from the \code{Species} slot of the first object in the list.
}
