% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovs.R
\name{getCovs}
\alias{getCovs}
\title{getCovs}
\usage{
getCovs(object, points = NA, visits = NA, reps = NA, times = NA,
  years = NA, months = NA, days = NA, lat = NA, long = NA,
  temp = NA, humid = NA, sky = NA, wind = NA, observer = NA,
  output = "dataframe")
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a list of such objects.}

\item{points}{A character vector or point names. Only visits to these points will be returned.}

\item{visits}{A length 1 numeric vector, defaults to NA. Returns data only from the indicated visits.}

\item{reps}{A length 1 numeric vector,d deflatus to NA, Returns only data form points and years where the point has been visited at least \code{reps} times in the year.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been vistied is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{years}{A numeric vector. Returns data only from points where the years the point was visited  matches one of the values in \code{years} The year a visit takes place is determined by the \code{Year} column in the \code{visits} slot which is dervied from the imformation in the \code{EventDate} column.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}
}
\description{
Returns the contents of the \code{Visits} and \code{Points} slots of an NCRNbirds object. The returned data can be filtered to meet various criteria.
}
\details{
This function returns point data either from a single NCRNbirds object or a list of such objects. The default output is a\code{data.frame}. However, if \code{object} is a list and \code{output} is "list" then a list of \code{data.frame}s will be returned. The name of each element in this list will correspond to the \code{ParkCode} in each NCRNbirds object.
}
