% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesign.R
\name{getDesign}
\alias{getDesign}
\title{getDesign}
\usage{
getDesign(object, info)
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a list of such objects.}

\item{info}{A length one chararcter vector. Indicates which aspect of the study design should be returned.
\describe{
\item{"visits"}{The number of visits typcially made to a point during a monitoring season. Retrieved from the \code{VisitNumber} slot.}
\item{"bands"}{The distance bands used for the study. Retrieved from the \code{Bands} slot.}
\item{"intervals"}{The time intervals used for the study. Retrieved from the \code{VisitNumber} slot.}
}}
}
\description{
Returns informaiton on the study design stored in an \code{NCRNbirds} object.
}
\details{
This function returns information about the study design used to collect the data. The type of information is 
determined by the \code{info} argument. When "visits" is selected the output will be a length one numeric vector, the other 
two options will return a\code{data.frame}. If the input is a \code{list} of \code{NCRNbird} objects, the the output will be eithe a 
list of vectors or a list of \code{date.frame}s depending on the \code{info} argument.
}
