% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGuilds.R
\name{getGuilds}
\alias{getGuilds}
\title{getGuilds}
\usage{
getGuilds(object, AOU = NA, categories = NA, output = "dataframe")
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a list of such objects.}

\item{AOU,}{an optional character vector of AOU codes. Each code should be in quotes. Only data from those birds will be returned.}

\item{categories}{A character vector of guild categories. Only data for those categories will be returned. Options are:
\describe{
\item{"Migratory"}{Migratory status of the species}
\item{"Brood.Numbers"}{Is the speices single or double brooded}
\item{"Primary.Habitat"}{Is the species a forest generalist or an interior forest obligate?}
\item{"Exotic"}{Is the species exotic or native?}
\item{"Nest.Placement"}{Where is the species' nest found? Possibilities include canopy, shrub, open ground nester etc.}
\item{"Foraging.Behavior"}{What tyoe of foraging behavior does the species show? Possibilites include wate plunger, bark prober,upper canopy forager, etc.}
\item{"Tropic.Level"}{What does the species feed on? Can be omnivore, carnivore, insectivore etc.}
\item{"Nest.Height"}{How high does the species nest? High, low or highly variable.}
\item{"Pred_Para_Desc"}{Is the species a nest predator or brood parasite.Yes or no.}
}}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function. When a list of \code{NCRNbirds} object is used, and the output is "dataframe" then a each unique conbination of species and guidls will only be listed once in the resulting dataframe. If some species have ifferent guild assingments in the different \code{NCRNbirds} objects in the list, then those species will be listed twice which could have unintended consiquences}
}
\description{
Returns the guild data from the \code{Species} slot of an NCRNbirds object. The returned data can be filtered to meet various criteria.
}
\details{
This function returns data on bird guilds from an \code{NPSForVeg} object or a list of such objects. This data is stored in the \code{Species} slot.
}

