% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkData.R
\name{getParkData}
\alias{getParkData}
\title{getParkData}
\usage{
getParkData(Dir, UnitCode)
}
\arguments{
\item{Dir}{The directory where the data is found. You should omit the trailing slash ("/") in the directory name.}

\item{UnitCode}{A character string (e.g., NCRN) to retrieve and assemble data for a desired park unit or network.}
}
\value{
Returns a list of 13 \code{NCRNbirds} objects, one for each park.
}
\description{
This function imports data from the standard Networks or Park Units as .csv files and saves it as \code{NCRNbirds} objects. The required .csv files are: Points, Visits, FieldData and BirdGuildAssignments.
}
