% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getParkNames.R
\name{getParkNames}
\alias{getParkNames}
\title{getNParkNames}
\usage{
getParkNames(object, name.class = "short")
}
\arguments{
\item{object}{Either an NPSbirds object or a list of such objects.}

\item{name.class}{Type of name to return. One of three options, in quotes.
\describe{
\item{"short"}{ The default. Returns the short name of the park}
\item{"long"}{Returns the long name of the park}
\item{"code"}{Returns the park code}
}}
}
\value{
A character vector with one or more park names.
}
\description{
Retreives park names from an NCRNbirds object or a list of such objects.
}

