% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCountUMF.R
\name{getPCountUMF}
\alias{getPCountUMF}
\title{getPCountUMF}
\usage{
getPCountUMF(object, points = NA, AOU = NA, years = NA, times = NA,
  band = 1, visits = NA, day = NA, observer = NA, temp = NA,
  humid = NA, sky = NA, mixture = "P", ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{times}{A numeric vector of length 1 passed on to \code{\link{CountXVisit}} and from there to \code{link{getVisits}}. Returns only data from points where the number of years that a point has been visited is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{band}{A numeirc vector. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}

\item{visits}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{...}{Additonal arguments passed to \code{\link{getPCountUMF}}.}
}
\description{
Creates an umarkedFrame object for fitting N-mixture models to count data with the \code{\link[unmarked]{pcount}} function (Fiske et al. 2011).
}
\details{
This function is used to create an unmarkedFrame object for inputing data for abundance estimation using the \code{\link[unmarked]{pcount}}function (Fiske et al. 2011).
}
\references{
Kery, M. Royle, J.A. and Schmid. 2005. Modeling avian abundance from replicated counts using binomail mixutre models. Ecological Applications. 15: 1450-1461.

Royle, J. A.  2004. N-Mixture models for estimating population size from spatially replicated counts. Biometrics 60: 108-115.

Fiske, I., R. Chandler, and others. 2011. Unmarked: An R package for fitting hierarchical models of wildlife occurrence and abundance. Journal of Statistical Software 43:1-23.
}
