% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdatedSpeciesInfo.R
\name{getUpdatedSpeciesInfo}
\alias{getUpdatedSpeciesInfo}
\title{getUpdatedSpeciesInfo}
\usage{
getUpdatedSpeciesInfo(Dir, ...)
}
\arguments{
\item{Dir}{A filepath directory to where the data resides.}
}
\description{
Gathers and saves up-to-date taxonomic and conservation status information on species to be used in updating species names and 4-letter codes within data, and generating tables.
}
\details{
This function navigates to existing URLs on the internet to download up-to-date taxonomic and conservation status information on species from Breeding Bird Survey data \link{ftp://ftpext.usgs.gov}, The Institute of Bird Populations \link{https://www.birdpop.org}, and Partners in Flight Bird Conservancy \link{http://pif.birdconservancy.org}. These data are then compiled and saved as a .csv file for future use in data anlayses and presentation.
}
