% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCNNAPbirds.R
\name{importCNNAPbirds}
\alias{importCNNAPbirds}
\title{importCNNAPbirds}
\usage{
importCNNAPbirds(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found. You should omit the trailing slash ("/") in the directory name.}
}
\value{
Returns 1 \code{NCRNbirds} object.
}
\description{
This function imports data from the standard CNNAP .csv files and saves it as \code{NCRNbirds} objects. 
The required .csv files are: Points, Visits, FieldData, CNNAPbands, CNNAPintervals, BirdSpecies, and BirdGuildAssignments.
}
