% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMIDNbirds.R
\name{importMIDNbirds}
\alias{importMIDNbirds}
\title{importMIDNbirds}
\usage{
importMIDNbirds(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found. You should omit the trailing slash ("/") in the directory name.}
}
\value{
Returns a list of 6 \code{NCRNbirds} objects, one for each park.
}
\description{
This function imports data from the standard MIDN .csv files and saves it as \code{NCRNbirds} objects. 
The required .csv files are: Points, Visits, FieldData MIDNbands, MIDNintervals, BirdSpecies, and BirdGuildAssignments.
}
