% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNCCNbirds.R
\name{importNCCNbirds}
\alias{importNCCNbirds}
\title{importNCCNbirds}
\usage{
importNCCNbirds(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found. You should omit the trailing slash ("/") in the directory name.}
}
\value{
Returns a list of 5 \code{NCRNbirds} objects, one for each park.
}
\description{
This function imports data from the standard NCCN .csv files and saves it as \code{NCRNbirds} objects. 
The required .csv files are: Points, Visits, FieldData NCCNbands, NCCNintervals, BirdSpecies, and BirdGuildAssignments.
}
