% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapOccupancy.R
\name{mapOccupancy}
\alias{mapOccupancy}
\title{mapOccupancy}
\usage{
mapOccupancy(object, points = NA, AOU = NA, years = NA, times = NA,
  band = 1, visits = NA, mixture = "P", site.covs = "Point_Name",
  obs.covs = NA, Table = TRUE, Figure = TRUE, color = NA,
  APIkey = NA, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{times}{A numeric vector of length 1 passed on to \code{\link{CountXVisit}} and from there to \code{\link{getVisits}}. Returns only data from points where the number of years that a point has been visited is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{band}{A numeirc vector. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}

\item{visits}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{mixture}{Defaults to "P". One of either "P", "NB", or "ZIP". Indicates which latent abundace distibution to use, either Poisson, Negative Binomial or Zero Inflated Poinsson.  Passed on to the \code{mixture} argumnet of \code{pcount} in the \code{unmarked} package.}

\item{site.covs}{A list of desired site-level covariates that can be included within the abundance estimation process of the N-mixture model.}

\item{obs.covs}{A list of desired observation-level covariates that can be included within the detection probability estimation process of the N-mixture model.}

\item{Table}{Logical (default=TRUE). Generates table to display and save, if desired.}

\item{Figure}{Logical (default=TRUE). Generates a figure to display and save, if desired.}

\item{color}{A character string naming a valid color for figures. Default is grayscale.}

\item{APIkey}{A character string of a Google API key used for generating maps with Google basemaps. If no APIkey is provided, basemaps default to stamen map.}

\item{...}{Additonal arguments passed to \code{\link{CountXVisit}} and from there to \code{\link{getBirds}}.}
}
\description{
Generates map of probability of site occupancy estimates at the point-level that pool data among years and control for detection probability.
}
\details{
This function generates a map of point-level abundance estimates following estimating abundance by fitting a N-mixture model (Royal 2004, Kery et al. 2005) to data in an NCRNbirds object or a list of such objects. The data is first extracted from the object(s) using the \code{\link{CountXVisit}} and \code{\link{CovsXVisit}} funcitons. The data is then fed to the \code{\link[unmarked]{pcount}} function and the resutls of the analysis are returned.
}
\references{
Kery, M. Royle, J.A. and Schmid. 2005. Modeling avian abundance from replicated counts using binomail mixutre models. Ecological Applications. 15: 1450-1461.

Royle, J. A.  2004. N-Mixture models for estimating population size from spatially replicated counts. Biometrics 60: 108-115.
}
