% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchParkCodes.R
\name{matchParkCodes}
\alias{matchParkCodes}
\title{matchParkCodes}
\usage{
matchParkCodes(object, parkNames, name.class = "short")
}
\arguments{
\item{object}{Either an NPSbirds object or a list of such objects.}

\item{parkNames}{A vector of Park codes}

\item{name.class}{Type of name to return. One of three options, in quotes.
\describe{
\item{"short"}{ The default. Returns the short name of the park}
\item{"long"}{Returns the long name of the park}
\item{"code"}{Returns the park code}
}}
}
\value{
A character vector with one or more park names.
}
\description{
Matches a string of park codes with the corresponding park names.
}
