% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nmixBirds.R
\name{nmixBirds}
\alias{nmixBirds}
\title{nmixBirds}
\usage{
nmixBirds(object, points = NA, AOU = NA, years = NA, times = NA,
  band = 1, visits = c(1, 2), mixture = "P", ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{times}{}

\item{visits,}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{mixture}{Defaults to "P". One of either "P", "NB", or "ZIP". Indicates which latent abundace distibution to use, either Poisson, Negative Binomial or Zero Inflated Poinsson.  Passed on to the \code{mixture} argumnet of \code{pcount} in the \code{unmarked} package.}

\item{...}{Additonal arguments passed to \code{\link{CountXVisit}} and from there to \code{\link{getBirds}}.}

\item{band.}{A numeirc vector. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}
}
\description{
Peforms trend analysis on bird data using N-mixture models (Royal 2004) form the unmarked package.
}
\details{
This funciton simplifed the process for fitting a N-mixture model (Royal 2004, Kery et al. 2005) to data in an NCRNbirds object or a list of such objects. The data is first extracted from the object(s) using the \code{\link{CountXVisit}} funciton. The data is then fed to the \code{\link{[unmarked]pcount}} function and the resutls of the analysis are returned.
}
\references{
Kery, M. Royle, J.A. and Schmid. 2005. Modeling avian abundance from replicated counts using binomail mixutre models. Ecological Applications. 15: 1450-1461.

Royle, J. A.  2004. N-Mixture models for estimating population size from spatially replicated counts. Biometrics 60: 108-115.
}

