% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmixBirds_zl.R
\name{nmixBirds_zl}
\alias{nmixBirds_zl}
\title{nmixBirds_zl}
\usage{
nmixBirds_zl(object, points = NA, AOU = NA, years = NA, times = NA,
  band = 1, visits = c(1, 2), site.covs = NA, obs.covs = NA,
  mixture = "P", ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{times}{A numeric vector of length 1 passed on to \code{\link{CountXVisit}} and from there to \code{\link{getVisits}}. Returns only data from points where the number of years that a point has been visited is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{band}{A numeirc vector. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}

\item{visits}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{site.covs}{A list of desired site-level covariates that can be included within the abundance estimation process of the N-mixture model.}

\item{obs.covs}{A list of desired observation-level covariates that can be included within the detection probability estimation process of the N-mixture model.}

\item{mixture}{Defaults to "P". One of either "P", "NB", or "ZIP". Indicates which latent abundace distibution to use, either Poisson, Negative Binomial or Zero Inflated Poisson.  Passed on to the \code{mixture} argumnet of \code{pcount} in the \code{unmarked} package.}

\item{...}{Additonal arguments passed to \code{\link{CountXVisit}}.}
}
\description{
Fits N-mixture models (Royal 2004) to count data using the pcount function (Fiske et al. 2011).
}
\details{
This function simplifed the process for fitting a N-mixture model (Royal 2004, Kery et al. 2005) to data in an NCRNbirds object or a list of such objects. The data is first extracted from the object(s) using the \code{\link{CountXVisit}} funciton. The data is then fed to the \code{\link[unmarked]{pcount}} function and the resutls of the analysis are returned.
}
\references{
Kery, M. Royle, J.A. and Schmid. 2005. Modeling avian abundance from replicated counts using binomail mixutre models. Ecological Applications. 15: 1450-1461.

Royle, J. A.  2004. N-Mixture models for estimating population size from spatially replicated counts. Biometrics 60: 108-115.

Fiske, I., R. Chandler, and others. 2011. Unmarked: An R package for fitting hierarchical models of wildlife occurrence and abundance. Journal of Statistical Software 43:1-23.
}
