% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParkTable.R
\name{ParkTable}
\alias{ParkTable}
\title{ParkTable}
\usage{
ParkTable(object, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{...}{Additonal arguments passed to \code{\link{ParkTable}}.}
}
\description{
Create a table of park units within a given network.
}
