% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointsTable.R
\name{PointsTable}
\alias{PointsTable}
\title{PointsTable}
\usage{
PointsTable(object, years = NA, visits = NA, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{visits}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{...}{Additonal arguments passed to \code{\link{PointsTable}}.}
}
\description{
Create a table of the number of sampling points at network- and park unit-levels among years.
}
\details{
This function generates a table tallying the number of sampling locations for an entire network and park units within each year.
}
