% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointsXVisitsTable.R
\name{PointsXVisitsTable}
\alias{PointsXVisitsTable}
\title{PointsXVisitsTable}
\usage{
PointsXVisitsTable(object, years = NA, visits = NA, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{years}{A vector of number. will return only data from the indicated years.}

\item{visits}{A length 1 numeric vector. Will only use data from points with that number of visits.}

\item{...}{Additonal arguments passed to \code{\link{PointsXVisitsTable}}.}
}
\description{
Create a table of the max number of visits to each sampling point in each year. NAs indicate the point was not sampled.
}
\details{
This function generates a table with the maximum number of times each point was visited in each year. NAs indicate points that were not sampled in given years.
}
