% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richnessPlot.R
\name{richnessPlot}
\alias{richnessPlot}
\title{richnessPlot}
\usage{
richnessPlot(object, years = NA, points = NA, plot_title = NA,
  point_num = NA, output = "total", ...)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{birdRichness()}.}

\item{years}{A nuemric vector. Indicates which years should be graphed.}

\item{points}{A character vector of point names. Only these points will be used.}

\item{plot_title}{Optional,  A title for the plot.}

\item{point_num}{An optional numeric vector indicating the number of points sampled each year. If \code{object} is a \code{NCRNbirds} object
or a \code{list} of such objects, then this will be calculated automatically. If \code{object} is a \code{data.frame} than this can be
provided by the user.}

\item{output}{Either "total" (the default) or "list". Only used when \code{obejct} is a \code{list}.}

\item{...}{Additional arguments passed to \code{\link{birdRichness}}}
}
\description{
Plots species richness over time.
}
\details{
This function produces a graph of species richness over time. It does this by using the output of the \code{\link{birdRichness}}
function. The data is then passed on to ggplot2 for graphing.
}
