% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesRichnessByParkFig.R
\name{SpeciesRichnessByParkFig}
\alias{SpeciesRichnessByParkFig}
\title{SpeciesRichnessByParkFig}
\usage{
SpeciesRichnessByParkFig(object, Dir = NA, color1 = NA, color2 = NA,
  ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{color1}{User-defined color for stacked bar chart.}

\item{color2}{User-defined color for stacked bar chart.}

\item{...}{Additonal arguments.}
}
\description{
Generates a stacked bar chart of species richness for all unique species and species of conservation concern (i.e., Partners in Flight watchlist species) at the network and park unit levels.
}
\details{
This function generates the maximum number of unique species at network and park unit levels across years, and presents these species richness values as a stacked bar chart with user-defined colors.
}
