% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesRichnessByParkTable.R
\name{SpeciesRichnessByParkTable}
\alias{SpeciesRichnessByParkTable}
\title{SpeciesRichnessByParkTable}
\usage{
SpeciesRichnessByParkTable(object, Dir, PIFspecies = TRUE, color = NA,
  Figure = TRUE, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{Dir}{A filepath directory to where the data resides.}

\item{PIFspecies}{Logical with default TRUE. Determines whether to present independent species richness table and heatmap for Partners in Flight (PIF) watchlist species.}

\item{color}{A character string naming a valid color for figures. Default is grayscale.}

\item{...}{Additonal arguments passed to \code{\link{SpeciesRichnessByParkTable}}.}

\item{Allspecies}{Logical with default TRUE. Determines whether to present independent species richness table and heatmap for all unique species.}
}
\description{
Genearates tables and heatmaps showing species richness at park unit and network levels among years.
}
\details{
This function generates tables (and heatmaps) containing network- and park-level species richness across years.
}
