% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesRichnessByPointTable.R
\name{SpeciesRichnessByPointTable}
\alias{SpeciesRichnessByPointTable}
\title{SpeciesRichnessByPointTable}
\usage{
SpeciesRichnessByPointTable(object, Dir, PIFspecies = TRUE, color = NA,
  Figure = TRUE, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{Dir}{A filepath directory to where the data resides.}

\item{PIFspecies}{Logical with default TRUE. Determines whether to present independent species richness table and heatmap for Partners in Flight (PIF) watchlist species.}

\item{color}{A character string naming a valid color for figures. Default is grayscale.}

\item{...}{Additonal arguments passed to \code{\link{SpeciesRichnessByPointTable}}.}

\item{Allspecies}{Logical with default TRUE. Determines whether to present independent species richness table and heatmap for all unique species.}
}
\description{
Generates a table and exports a .csv file, and generates a heatmap figure of point-level species richness.
}
\details{
This function generates tables (and heatmaps) containing network- and park-level species richness at each sampling point and year.
}
