% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesRichnessMap.R
\name{SpeciesRichnessMap}
\alias{SpeciesRichnessMap}
\title{SpeciesRichnessMap}
\usage{
SpeciesRichnessMap(object, Dir, PIFspecies = TRUE, color = NA,
  APIkey = NA, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{Dir}{A filepath directory to where the data resides.}

\item{PIFspecies}{Logical with default TRUE. Determines whether to present independent species richness map for Partners in Flight (PIF) watchlist species.}

\item{color}{A character string naming a valid color for figures. Default is grayscale.}

\item{APIkey}{A character string of a Google API key used for generating maps with Google basemaps. If no APIkey is provided, basemaps default to stamen map.}

\item{...}{Additonal arguments passed to \code{\link{SpeciesRichnessMap}}.}
}
\description{
Generates a map of point-level species richness.
}
\details{
This function determines the maximum count of unique species at each sampling point, and generates maps for either or both all unique species and PIF watchlist species across sampling years.
}
