% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesTable.R
\name{SpeciesTable}
\alias{SpeciesTable}
\title{SpeciesTable}
\usage{
SpeciesTable(object, Dir, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{...}{Additonal arguments passed to \code{\link{SpeciesTable}}.}
}
\description{
Generate a table of species with corresponding taxonomic and conservation status information.
}
\details{
This function generates a table (and .csv file to save, if desired) of all unique species within the data provided by the object. The table includes up-to-date information on common names, scientific names, and 4-letter alpha codes from \code{\link{getUpdatedSpeciesInfo}} function, as well as conservation status information for each species.
}
