% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeEffort.R
\name{summarizeEffort}
\alias{summarizeEffort}
\title{summarizeEffort}
\usage{
summarizeEffort(
  object,
  byYear = T,
  byPoint = F,
  effort = "points",
  wide = F,
  output = "dataframe",
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a \code{list} of such objects.}

\item{byYear}{Defaults to \code{TRUE}. Indicates if data should be separated by year (\code{TRUE}) or combined (\code{FALSE})}

\item{byPoint}{Defaults to \code{FALSE}. Indicates if data should be separated by point (\code{TRUE}) or combined (\code{FALSE})}

\item{effort}{Indicates if effort should be measured by "points" (# of points visited) or "visits" (total # of visits).}

\item{wide}{Defaults to \code{FALSE}. If \code{TRUE} and \code{wide} is \code{TRUE} then there will be a column for each year and row for each park 
or point. Otherwise there will be a single \code{Year} column to indicate the year.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}

\item{...}{Additional arguments passed to \code{getVisits}}
}
\description{
Produces a table indicating how many visits were made to parks and points.
}
\details{
This produces a summary of the amoutn of effort put into monitoring either measured as the number points visited or the number of 
visits made to the points. The data on effort is retrieved from the \code{Visits} slot of the \code{NCRNbirds} object(s). The output can either be 
in long format - suitable for further use in a map or graph, or in wide format which is easier for viewing. Any argument which is valid for 
\code{getVisits} is also valid for this function.
}
