% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTrend.R
\name{summarizeTrend}
\alias{summarizeTrend}
\title{summarizeTrend}
\usage{
summarizeTrend(object, tabletype = "state", ...)
}
\arguments{
\item{object}{An unamked frame object such as those created by \code{makeUMF} or a \code{list} of such frames.}
}
\value{
A table which summarizes the trend analysis results. Currently fits occupancy and n-mixture models.
}
\description{
Analyzes an unmarkedFrame object using methods from the unmarked package.
}
\details{
This function simplifies the process of extracting trends over tiem from unmarked models (Royal 2004, Kery et al. 2005) generated
by the \code{\link{unmarkedBirds}} command. The code currenlty assumes you have an occupancy model fit with the trend Year as a factor.
}
