% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsPlot.R
\name{trendsPlot}
\alias{trendsPlot}
\title{trendsPlot}
\usage{
trendsPlot(
  object,
  trendtype,
  estimate = "psi",
  est_symbol = "line",
  est_color = "dark green",
  est_shape = 19,
  obs_color = "black",
  obs_shape = 17,
  error_symbol = "ribbon",
  plot_title = NA,
  legend = T,
  plot = TRUE
)
}
\arguments{
\item{object}{An unmarked fit object with an yearly trend such as those produced by \code{\link{unmarkedBirds}}, a \code{data.frame} produced by 
\code{\link{summarizeTrend}}, or a \code{list} of such objects.}

\item{trendtype}{Either "occu" or "pcount". Indicates if this is from an occupancy model or an n-mixture model. Only required if \code{object} is a data.frame.}

\item{estimate}{Either "psi", the default, or "z". Inidicates the type of esimate you wish to display.}

\item{est_symbol}{Either "point" or "line". Indicates how you wish to display the estimate.}

\item{est_color}{The color of the estimate.}

\item{est_shape}{If \code{est_symbol} is "point" this will control the shape of the points. Default is a filled circle.}

\item{obs_color}{The color of the observed data.}

\item{obs_shape}{This controls the shape of the observed points. Default is a filled triangle.}

\item{error_symbol}{This controls how the 95\% CIs are displayed. Either "bar" for error bars or "ribbon" for a lightly shaded region.}

\item{plot_title}{Optional,  A title for the plot.}

\item{legend}{If \code{TRUE} will include a legend in the plot.}

\item{plot}{Logical. Return plot \code{TRUE} (default) or data.frame \code{FALSE}.}
}
\description{
Makes a plot of trends in bird populations
}
\details{
This function produces plot of trends based on the output of \code{\link{summarizeTrend}}. If the \code{object} is the output of 
\code{\link{unmarkedBirds}} then the function will automatically detect if the trend is based on an occupancy or n-mixture model. If the \code{object}
is a \code{data.frame} this has to be specified using the \code{trendtype} agrument. 

The plot will display both the naive values and the model estimates. The type of estimate is controlled by the \code{esitmate} argument. If this 
is "psi", then the model estimates will be the model coefficients back transformed to occupancy or abundance. This is the estimated 
value for  a site given its site. The  other option,"z", takes into account not only site covariates but also the  visit covaraites and 
actual detection history for each site.  The estiamte will also display  upper and lower limits for a 95\% CI. 
 
Several arguments control the graphical display. The naive values are always points, but their color and shape can be selected by the user. Estiamated 
values can also be displayed and customized as points, or can be a line instead. Error in the estimates can either be shown as yearly error bars or as
a "ribbon" -  a shaded area around the estimate. The color and shape arguments accept standard \code{ggplot2} options.
}
