% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateSpeciesInfo.R
\name{UpdateSpeciesInfo}
\alias{UpdateSpeciesInfo}
\title{UpdateSpeciesInfo}
\usage{
UpdateSpeciesInfo(object, Dir, ...)
}
\arguments{
\item{object}{An NCRNbirds object or a list of such objects.}

\item{Dir}{A filepath directory to where the data resides.}

\item{...}{Additonal arguments passed to \code{\link{UpdateSpeciesInfo}}.}
}
\description{
Update count data by finding and replacing inaccurate taxonomic names and species codes.
}
\details{
This function queries the user to input desired list of species common names (as currently exisiting in the data) for updating and reconciliation with updated species taxonomic information and conservation status gathered from \code{\link{getUpdatedSpeciesInfo}}. It saves new data, and archives original data.
}
