% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotRichness.R
\name{barplotRichness}
\alias{barplotRichness}
\title{barplotRichness}
\usage{
barplotRichness(
  object,
  byPark = F,
  byYear = F,
  byPoint = F,
  color = "dark green",
  plot_title = NA,
  scale = T,
  labels = T,
  output = "total",
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{\link{birdRichness}}.}

\item{byPark}{}

\item{byPoint}{Defaults to \code{FALSE}. If \code{TRUE} heatamap will have one line for each point, otherwise will have one line for each park}

\item{color}{A color that will correspond to the maximum species richness}

\item{plot_title}{Optional,  A title for the plot.}

\item{scale}{If \code{TRUE} will include a scale to the right of the heatmap}

\item{labels}{If\code{TRUE} will label species richness values on the heatmap.}

\item{output}{Either "total" (the default) or "list". Only used when \code{object} is a \code{list}}

\item{plot}{Logical. Return plot \code{TRUE} (default) or data.frame \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link{birdRichness}}}
}
\description{
Makes a barplot of species richness
}
\details{
This function produces a heatmap of species richness at the park or point level across years. It does this by using the 
output of the \code{\link{birdRichness}} function. The data is then passed on to ggplot2 for graphing.
}
