% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountXVisit.R
\name{CountXVisit}
\alias{CountXVisit}
\title{CountXVisit}
\usage{
CountXVisit(
  object,
  parks = NA,
  points = NA,
  AOU = NA,
  years = NA,
  times = NA,
  band = 1,
  visits = NA,
  max = F,
  site = NA,
  dist = NA,
  wind = NA,
  sky = NA,
  output = "dataframe",
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a \code{list} of such objects.}

\item{parks}{A character vector of park codes. Only visits within these parks will be returned.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.}

\item{years}{A vector of numbers. Will return only data from the indicated years.}

\item{times}{A numeric vector of length 1 passed on to \code{link{getVisits}}. Returns only data from points where the number of years that a point has 
been visited is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{band}{A numeric vector. Defaults to 1. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}

\item{visits}{The visits that will be used for the matrix. Defaults to \code{NA}. See Details below.}

\item{max}{Logical, defaults to \code{FALSE}. If \code{TRUE} then the matrix will also include a single column with the maximum detections seen at each 
point across all the visits specified in the \code{visits} argument.}

\item{site}{Character. Select sites in "Forest" or "Grassland" habitats. Defaults to selecting both site types.}

\item{dist}{Character. Filter data by disturbance code(s). See network documentation for list and meanings of codes.}

\item{wind}{Numeric. Filter data by wind code(s). See network documentation for list and meanings of codes.}

\item{sky}{Numeric. Filter data by sky condition code(s). See network documentation for list and meanings of codes.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}

\item{...}{Additional arguments passed to \code{getBirds}}
}
\description{
Produces a Count X Visit matrix for use in analyses
}
\details{
This produces a Count X Visit matrix for a \code{NCRNbirds} object or a \code{list} of such objects. Each row of the matrix
 will correspond to a different point in a different year. The columns of the matrix will be the park code, the point name, the year 
 visited, and a column of abundances of the indcated species at that visit. If multiple species are indicated in \code{AOU}, their 
 abundances will be totaled. If \code{max=T} then the maximum accross all visits will be returned as well as the indvidual visit totals.
 
 If \code{visits} is left as \code{NA} then the visits used will be 1 through the number of visits indicated in the \code{visits} slot. 
 Otherwise a numeric vectore e.g. c(1,2) can be used to select which visits are used.
}
