% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovsXVisit.R
\name{CovsXVisit}
\alias{CovsXVisit}
\title{CovsXVisit}
\usage{
CovsXVisit(
  object,
  covs,
  parks = NA,
  points = NA,
  years = NA,
  times = NA,
  visits = NA,
  reps = NA,
  output = "dataframe"
)
}
\arguments{
\item{object}{A \code{data.frame},  \code{NCRNbirds} object or a list of such objects. If the input is a \code{data.frame} in should be in the 
same format at the ouptut of \code{\link{getVisits}}}

\item{covs}{A character vector of one or more covariate names. Options inlcude:
\describe{
\item{"day"}{Provides the Julian day (aka ordinal day) for each visit.}
\item{"humidity"}{Provides the humidity measured for each visit.}
\item{"observer"}{Provides the observer for each visit.}
\item{"sky"}{Provides the sky condiiton (e.g. Clear, Cloudy, Fog etc.) for each visit.}
\item{"temp"}{Provides the temperature for each visit.}
\item{"decibel"}{Provides the backround noise level in decibels.}
\item{"time"}{Provides the time of day when each visit began. This is expressed in minutes since the start of the day.}
\item{"visit"}{Indicates if a visit occurred or not. Each visit column will either have the visit number if a visit occured or \code{NA}
if it did not.}
\item{"wind"}{Provides the wind level for each visit (e.g. Calm, Light breeze).}
}}

\item{parks}{A character vector of park codes. Only data from these parks will be returned.}

\item{points}{A character vector of point names. Only data from these points will be returned.}

\item{years}{A numeric vector. Returns data only from points where the years the point was visited matches one of the 
values in \code{years}. The year a visit takes place is determined by the \code{Year} column in the \code{visits} slot 
which is derived from the information in the \code{EventDate} column.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been vistied 
is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{visits}{A length 1 numeric vector, defaults to \code{NA}. Returns data only from the indicated visits.}

\item{reps}{A length 1 numeric vector,defaults to \code{NA}. Returns only data from points and years where the point has been visited 
at least \code{reps} times in the year.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}
}
\description{
Produces a Covariate X Visit matrix for use in analyses
}
\details{
This produces a Covariate(s) X Visit matrix for a \code{NCRNbirds} object or a \code{list} of such objects. Each row of the matrix
 will correspond to a different point in a different year. The columns of the matrix will be the park code, the point name, the year 
 visited, and a column for each covariate at each visit. These columns will be called "Day1", "Day2"  or  "Humidity1", "Humidity2" etc. to indicate
 the covariate and the visit.  
 
 If \code{visits} is left as \code{NA} then the visits used will be 1 through the number of visits indicated in the \code{visits} slot. 
 Otherwise a numeric vector e.g. c(1,2) can be used to select which visits are used. 
 
 If the input \code{object} is an \code{NCRNbirds} object or a list of such objects, the function will use the \code{\link{getVisits}} function to 
 retrieve the visit data. Currently if you supply a \code{data.frame} as the \code{object} the filtering by parks, points etc. will not occur, 
 all visits present in the input \code{data.frame} will be present it the ouput \code{data.frame}
}
