% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectsPlot.R
\name{detectsPlot}
\alias{detectsPlot}
\title{detectsPlot}
\usage{
detectsPlot(
  object,
  parks = NA,
  years = NA,
  points = NA,
  visits = NA,
  times = NA,
  max = F,
  plot_title = NA,
  point_num = NA,
  se = FALSE,
  add_line = TRUE,
  output = "total",
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{CountXVisit()}.}

\item{parks}{A character vector of park codes. Only visits within these parks will be returned.}

\item{years}{A numeric vector. Indicates which years should be graphed.}

\item{points}{A character vector of point names. Only these points will be used.}

\item{visits}{A length 1 numeric vector, defaults to NA. Returns data only from the indicated visits.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been vistied is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{max}{Logical. If \code{FALSE}, the default, then the annual mean will be calculated from each visit separately in each year. 
If \code{TRUE} then the annual mean will be calculated from the maximum no. detected across all visits in each year.}

\item{plot_title}{Optional,  A title for the plot.}

\item{point_num}{An optional list of numeric vector indicating the number of points sampled each visit of each year. Each visit is a separte elemnet in the list. If \code{object} is a \code{NCRNbirds} object
or a \code{list} of such objects, then this will be calculated automatically. If \code{object} is a \code{data.frame} than this can be
provided by the user.}

\item{se}{Logical. Add Standard error to the plot. \code{TRUE} or \code{FALSE}, defaults to \code{FALSE}}

\item{add_line}{Logical. Connects points in plot when \code{TRUE}. Defaults to \code{TRUE}.}

\item{output}{Either "total" (the default) or "list". Only used when \code{object} is a \code{list}.}

\item{...}{Additional arguments passed to \code{\link{CountXVisit}}}
}
\description{
Plots bird detections (relative abundance) over time.
}
\details{
This function produces a graph of species detections over time. It does this by using the output of the \code{\link{CountXVisit}}
function. The data is then passed on to ggplot2 for graphing.
}
