% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAKNData.R
\name{getAKNData}
\alias{getAKNData}
\title{getAKNData}
\usage{
getAKNData(Dir, import = FALSE)
}
\arguments{
\item{Dir}{The directory where the data is found. You should omit the trailing slash ("/") in the directory name.}

\item{import}{Logical. If \code{TRUE} will import data as "NETN" NCRNbirds object.}
}
\description{
Takes NETN's AKN point count survey data download and updates Visits and FieldData files for NCRNbirds package import. Also has option to import into workspace as NCRNbirds object ("NETN").
}
