% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVisits.R
\name{getVisits}
\alias{getVisits}
\title{getVisits}
\usage{
getVisits(object, parks = NA, times = NA, years = NA, points = NA,
  visits = NA, reps = NA, output = "dataframe")
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a list of such objects.}

\item{parks}{A character vector of park codes. Only visits within these parks will be returned.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been vistied is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{years}{A numeric vector. Returns data only from points where the years the point was visited  matches one of the values in \code{years} The year a visit takes place is determined by the \code{Year} column in the \code{visits} slot which is dervied from the imformation in the \code{EventDate} column.}

\item{points}{A character vector of point names. Only visits to these points will be returned.}

\item{visits}{A length 1 numeric vector, defaults to NA. Returns data only from the indicated visits.}

\item{reps}{A length 1 numeric vector,d deflatus to NA, Returns only data form points and years where the point has been visited at least \code{reps} times in the year.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}
}
\description{
Returns the contents of the \code{Visits} slot of an NCRNbirds object. The returned data can be filtered to meet various criteria.
}
\details{
This function returns point data either from a single NCRNbirds object or a list of such objects. The default output is a\code{data.frame}. However, if \code{object} is a list and \code{output} is "list" then a list of \code{data.frame}s will be returned. The name of each element in this list will correspond to the \code{ParkCode} in each NCRNbirds object.
}
