% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMIDNbirds.R
\name{importMIDNbirds}
\alias{importMIDNbirds}
\title{importMIDNbirds}
\usage{
importMIDNbirds(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found}
}
\value{
Returns a list of 11 \code{NCRNbirds} objects, one for each park.
}
\description{
This function imports data from the standard NCRN .csv files and saves it as \code{NCRNbirds} objects. The required .csv files are: Points, Visits, FieldData and BirdGuildAssignments.
}
