% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapUnmarked.R
\name{mapUnmarked}
\alias{mapUnmarked}
\title{mapUnmarked}
\usage{
mapUnmarked(
  object,
  model,
  estimate = "psi",
  years = NA,
  points = NA,
  palette = "BuGn",
  maptype = "basic",
  title = "Model Estimates",
  ...
)
}
\arguments{
\item{object}{Either an object of class \code{NCRNbirds} or a list of such objects.}

\item{model}{An unmarked fit object such as those created by \code{\link{unmarkedBirds}}.}

\item{estimate}{Either "psi", the default, or "z". Inidicates the type of esimate you wish to display.}

\item{years}{A vector of a single numeric value indicating the year of the estimates you wish to map. Do not include multiple years
as the map shows preictios for a particular year.}

\item{points}{A character vector of point names. \code{NA}, the default, will map all points.}

\item{palette}{Color pallete for the colors of the points. Defaults to "BuGn" (blue green) but will accept any RColorBrewer, viridis or custom palette.}

\item{maptype}{The type of base map to be used. See \code{\link{mapBirds}} for options.}

\item{title}{A character vector to be used as the legend title.}

\item{...}{Additional arguments passed to \code{\link{birdRichness}}.}
}
\description{
Produces an html map of fits for an unmarked fit.
}
\details{
This function prodcues a map showing model fits for unmarked fit objects for a given year, as produced by \code{\link{unmarkedBirds}}. In order to produce
the map the function requries you to specify which year's fits you are interestred in. The fits are then passed on to the \code{\link{mapBirds}} function.
}
\seealso{
\code{\link{unmarkedBirds}} \code{\link{mapBirds}}
}
