% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRNbirds_Class_def.R
\docType{class}
\name{NCRNbirds-class}
\alias{NCRNbirds-class}
\title{S4 Class Definition for NCRNbirds}
\description{
An S4 class that contains the data from bird monitoring at a set of monitoring points. This will typcially be a single park, but it could be any group of plots, including part of a park or a group of parks.
}
\section{Slots}{

\describe{
\item{\code{ParkCode}}{A short code to designate the park (or group of plots), typically an NPS 4 letter code. Stored as a length 1 character vector.}

\item{\code{ShortName}}{A short name for the park. Stored as a length 1 character vector.}

\item{\code{LongName}}{A long, formal name for the park. Stored as a length 1 character vector.}

\item{\code{Network}}{The code for the Inventory & Montoirng network (or other network) the park belongs to. Stored as a length 1 character vector.}

\item{\code{VisitNumber}}{A \code{numeric} vector of length 1. The number of visits that is typically made to each point. This serves as the default number of visits for other functions.}

\item{\code{Points}}{A \code{data.frame} with metadata for the monitoring points, such as names and lat/long.}

\item{\code{Visits}}{A \code{data.frame} with metadata for each sampling visit.}

\item{\code{Birds}}{A \code{data.frame} with the bird monitoring data. Each row is an observation of a number of individuals of a particular species at a particular point, visit, time interval and distance band.}

\item{\code{Species}}{A \code{data.frame} with metadata about the bird species, including AOU (American Ornithological Union) Code, Latin and common names, as well as guild assignments for the BCI calculations.}
}}

