% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richnessPlot.R
\name{richnessPlot}
\alias{richnessPlot}
\title{richnessPlot}
\usage{
richnessPlot(
  object,
  years = NA,
  points = NA,
  visits = NA,
  times = NA,
  plot_title = NA,
  point_num = NA,
  add_line = TRUE,
  output = "total",
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{birdRichness()}.}

\item{years}{A numeric vector. Indicates which years should be graphed.}

\item{points}{A character vector of point names. Only these points will be used.}

\item{visits}{A length 1 numeric vector, defaults to NA. Returns data only from the indicated visits.}

\item{times}{A numeric vector of length 1. Returns only data from points where the number of years that a point has been visited is greater 
or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{plot_title}{Optional,  A title for the plot.}

\item{point_num}{An optional numeric vector indicating the number of points sampled each year. If \code{object} is a \code{NCRNbirds} object
or a \code{list} of such objects, then this will be calculated automatically. If \code{object} is a \code{data.frame} than this can be
provided by the user.}

\item{add_line}{Logical. Connects points in plot when \code{TRUE}. Defaults to \code{TRUE}.}

\item{output}{Either "total" (the default) or "list". Only used when \code{object} is a \code{list}.}

\item{plot}{Logical. Return plot \code{TRUE} (default) or data.frame \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link{birdRichness}}}
}
\description{
Plots species richness over time.
}
\details{
This function produces a graph of species richness over time. It does this by using the output of the \code{\link{birdRichness}}
function. The data is then passed on to ggplot2 for graphing.
}
