% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTrend.R
\name{summarizeTrend}
\alias{summarizeTrend}
\title{summarizeTrend}
\usage{
summarizeTrend(object, ...)
}
\arguments{
\item{object}{An unmarked fit object such as those created by \code{\link{unmarkedBirds}} or a \code{list} of such fits.}
}
\value{
A table which summarizes the trend analysis results. Currently fits occupancy models with only year (factor or numeric) as a site covariate. 
If the input is a \code{list} of unmarked fits, then the function will return a \code{list} of table.s
}
\description{
Summarizes trends for an unmarkedFit object using methods from the unmarked package.
}
\details{
This function simplifies the process of extracting trends from unmarked models (Royal 2004, Kery et al. 2005) generated
by the \code{\link{unmarkedBirds}} function. 

The function will return a table, where each row is a different year. Columns include the year, the number of points visited., naive 
occupancy or abundance for each year, the model coefficient and standard error. The function also returns two sets of estimates of the true 
occupancy or abundance, by each year. The first, \emph{psi}, is the inverse model coefficient back transformed to occupancy / abundance and 
represents the estimated value for  a site given its site and visit covariates. The second estimate, \emph{z}, takes into account not only covariates
but also the actual detection history for each site.  Each estimate comes with  its own upper and lower limits for a 95\% CI. Note that as \emph{z} 
takes into account the actual detection history, its lower occupancy limit is always at least the naive occupancy, whereas \emph{psi} has no such 
lower boundary.
}
