% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumRelAbund.R
\name{SumRelAbund}
\alias{SumRelAbund}
\title{SumRelAbund}
\usage{
SumRelAbund(
  object,
  parks = NA,
  points = NA,
  AOU = NA,
  years = NA,
  times = NA,
  band = 1,
  visits = NA,
  CalcByYear = FALSE,
  max = TRUE,
  sort = FALSE,
  abund = 10,
  output = "dataframe",
  ...
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object or a \code{list} of such objects.}

\item{parks}{A character vector of park codes. Only visits within these parks will be returned.}

\item{points}{A character vector. The names of one or more points where the data was collected.}

\item{AOU}{A character vector. One or more AOU (American Onothological Union) codes of bird species.
Detections will be summed by each individual species.}

\item{years}{A vector of numbers. Will return only data from the indicated years.}

\item{times}{A numeric vector of length 1 passed on to \code{\link{getVisits}} and \code{\link{getBirds}}. Returns only data from points where the number of years that a point has been visited is greater or equal to the value of \code{times}. This is determined based on the data found in the \code{Visits} slot.}

\item{band}{A numeric vector. Defaults to 1. Only observations whose \code{Distance_id} field matches a value in \code{band} will be returned.}

\item{visits}{The visits that will be used for the matrix. Defautls to \code{NA}. See Details below,}

\item{CalcByYear}{Logical, if \code{TRUE}, will calculate mean detections across all visits per year. Defaults to \code{FALSE}, calculating per visit.}

\item{max}{Logical, if \code{TRUE} (default), relative abundance will be calculated from the maximum count among visits in a given year.}

\item{sort}{Logical, if \code{TRUE}, when multiple species are selected it will calculate and sort relative abundance per species. See \code{abund}.}

\item{abund}{Numeric, When \code{sort} = \code{TRUE}, used to provide a numeric value to select the most abundant species.
E.g., abund = 10 will return mean detections of the top 10 species. You can use the returned  \code{data.frame} to provide species AOU.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}

\item{...}{Additional arguments passed to \code{getChecklist}}
}
\description{
Produces a summary of raw detections by species for plotting and analysis.
}
\details{
Summarizes relative abundance by species (mean detections per point) for a \code{NCRNbirds} object or a \code{list} of such objects. 
 
 If \code{visits} is left as \code{NA} then the visits used will be 1 through the number of visits indicated in the \code{visits} slot. 
 Otherwise a numeric vectore e.g. c(1,2) can be used to select which visits are used. 
 
 \code{SumRelAbund} requires at least 2 monitoring points or at least 2 years of data to be specified, unless both \code{CalcByYear} and
  \code{sort} are \code{TRUE}. 
 
 If \code{sort} is \code{TRUE}, then data will be combined accross the visits, points and years indicated by the other arguments.
}
