% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsPlot.R
\name{trendsPlot}
\alias{trendsPlot}
\title{trendsPlot}
\usage{
trendsPlot(
  object,
  trendtype,
  estimate = "psi",
  est_symbol = "line",
  est_color = "dark green",
  est_shape = 19,
  obs_color = "black",
  obs_shape = 17,
  error_symbol = "ribbon",
  plot_title = NA,
  legend = T,
  plot = TRUE
)
}
\arguments{
\item{object}{An \code{NCRNbirds} object a \code{list} of such objects, or a \code{data.frame} like that produced by \code{\link{birdRichness}}.}

\item{trendtype}{Indicates if this is from an occupancy model "occu" or a n-mixture model. Only required if \code{object} is a data.frame.}

\item{est_symbol}{Either "point" or "line". Indicates how you wish to display the estimate.}

\item{est_color}{The color of the estimate.}

\item{est_shape}{If \code{est_symbol} is "point" this will control the shape of the points. Default is a filled circle.}

\item{obs_color}{The color of the observed data.}

\item{obs_shape}{This controls the shape of the observed points. Default is a filled triangle.}

\item{error_symbol}{This controls how the 95\% CIs are displayed. Either "bar" for error bars or "ribbon" for a lightly shaded region.}

\item{plot_title}{Optional,  A title for the plot.}

\item{legend}{If \code{TRUE} will include a legend in the plot.}

\item{plot}{Logical. Return plot \code{TRUE} (default) or data.frame \code{FALSE}.}

\item{estiamte}{Either "psi", the defualt or "z". Inidicate the type of esimate you wish to display.}
}
\description{
Makes a plot of trends in bird populations
}
\details{
This function produces a barplot of species richness by Park, Year or Point. It does this by using the 
output of the \code{\link{birdRichness}} function. The data is then passed on to ggplot2 for graphing. The user can chose to have \code{byPark=T}, 
\code{byYear=T} or \code{byPoint=T}, but should only pick one of the three. These can be combined with a guild category using \code{byGuild}. The 
\code{includeNA} argument indicates if bird species that are not categorized for a particular guild shoud be included.

Colors:

Colors of the bars are controlled by either \code{colors} or \code{palette} argumnts. If the barplot is not by guild then the color indicated by 
\code{colors} will be used. If \code{byGuild=T} then by default the \code{palette} arguement will be used. However, the \code{palette} arguement
can be overridden simply by giving \code{colors} a vector of colors with one color for each guild category shown in the barplot.If 
\code{includeNA=T} than than the first color will correspond to the NA value.
}
